+++ pkg/runtime/collections/array_iterator_test.go
@@ -1,0 +1,111 @@
+package collections_test
+
+import (
+	"github.com/MontFerret/ferret/pkg/runtime/collections"
+	"github.com/MontFerret/ferret/pkg/runtime/core"
+	"github.com/MontFerret/ferret/pkg/runtime/values"
+	. "github.com/smartystreets/goconvey/convey"
+	"testing"
+)
+
+func TestArrayIterator(t *testing.T) {
+	Convey("Should iterate over an array", t, func() {
+		arr := values.NewArrayWith(
+			values.NewInt(1),
+			values.NewInt(2),
+			values.NewInt(3),
+			values.NewInt(4),
+			values.NewInt(5),
+		)
+
+		iter := collections.NewArrayIterator(arr)
+
+		res := make([]core.Value, 0, arr.Length())
+
+		pos := 0
+
+		for iter.HasNext() {
+			item, key, err := iter.Next()
+
+			So(err, ShouldBeNil)
+			So(key.Unwrap(), ShouldEqual, pos)
+
+			res = append(res, item)
+
+			pos += 1
+		}
+
+		So(res, ShouldHaveLength, arr.Length())
+	})
+
+	Convey("Should iterate over an array in the same order", t, func() {
+		arr := values.NewArrayWith(
+			values.NewInt(1),
+			values.NewInt(2),
+			values.NewInt(3),
+			values.NewInt(4),
+			values.NewInt(5),
+		)
+
+		iter := collections.NewArrayIterator(arr)
+
+		res := make([]core.Value, 0, arr.Length())
+
+		for iter.HasNext() {
+			item, _, err := iter.Next()
+
+			So(err, ShouldBeNil)
+
+			res = append(res, item)
+		}
+
+		arr.ForEach(func(expected core.Value, idx int) bool {
+			actual := res[idx]
+
+			So(actual, ShouldEqual, expected)
+
+			return true
+		})
+	})
+
+	Convey("Should return an error when exhausted", t, func() {
+		arr := values.NewArrayWith(
+			values.NewInt(1),
+			values.NewInt(2),
+			values.NewInt(3),
+			values.NewInt(4),
+			values.NewInt(5),
+		)
+
+		iter := collections.NewArrayIterator(arr)
+
+		res := make([]core.Value, 0, arr.Length())
+
+		for iter.HasNext() {
+			item, _, err := iter.Next()
+
+			So(err, ShouldBeNil)
+
+			res = append(res, item)
+		}
+
+		item, _, err := iter.Next()
+
+		So(item, ShouldEqual, values.None)
+		So(err, ShouldBeError)
+	})
+
+	Convey("Should NOT iterate over an empty array", t, func() {
+		arr := values.NewArray(10)
+
+		iter := collections.NewArrayIterator(arr)
+
+		var iterated bool
+
+		for iter.HasNext() {
+			iterated = true
+		}
+
+		So(iterated, ShouldBeFalse)
+	})
+}
